Port JSON and CSV readers to datasource API.
